Stop using WidgetEngine as its merged into Widget.
